+++ website/public/sitemap-0.xml
@@ -1,8 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://yourdomain.com/p/dashboard</loc><lastmod>2025-04-14T17:11:10.855Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://yourdomain.com</loc><lastmod>2025-04-14T17:11:10.856Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://yourdomain.com/login</loc><lastmod>2025-04-14T17:11:10.856Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://yourdomain.com/privacy</loc><lastmod>2025-04-14T17:11:10.856Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://yourdomain.com/terms</loc><lastmod>2025-04-14T17:11:10.856Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://yourdomain.com/login</loc><lastmod>2025-06-07T19:04:03.158Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://yourdomain.com</loc><lastmod>2025-06-07T19:04:03.159Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://yourdomain.com/login/reset</loc><lastmod>2025-06-07T19:04:03.159Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://yourdomain.com/login/reset/callback</loc><lastmod>2025-06-07T19:04:03.159Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://yourdomain.com/login/success</loc><lastmod>2025-06-07T19:04:03.159Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://yourdomain.com/privacy</loc><lastmod>2025-06-07T19:04:03.159Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://yourdomain.com/login/verify</loc><lastmod>2025-06-07T19:04:03.159Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://yourdomain.com/terms</loc><lastmod>2025-06-07T19:04:03.159Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://yourdomain.com/login/verify/callback</loc><lastmod>2025-06-07T19:04:03.159Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://yourdomain.com/p/dashboard</loc><lastmod>2025-06-07T19:04:03.159Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
